backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart